chore: store, add comments

+++ core/map/map.ts
@@ -5,7 +5,7 @@ type StoreOptions = {
     owner: string;
     repo: string;
     path: string;
-    token: string;
+    token?: string;
     branch?: string;
 } | null;
 
@@ -20,25 +20,21 @@ export class Store<K extends string | number | symbol, V> {
         this.options = options;
     }
 
+    /**
+     * Adds a new element with a specified key and value to the Map. If an element with the same key already exists, the element will be updated.
+     * @param key
+     * @param value
+     * @param ttl
+     */
     set(key: K, value: V, ttl?: number): void {
         const expiry = ttl ? Date.now() + ttl : undefined;
         this.map.set(key, { value, expiry });
     }
 
-    async refresh() {
-        if (this.map.size === 0 && this.options) {
-            const map = await getMap<K, V>({
-                token: this.options.token,
-                owner: this.options.owner,
-                repo: this.options.repo,
-                path: this.options.path,
-                branch: this.options.branch,
-            });
-            if (!map) return;
-            this.map = map;
-        }
-    }
-
+    /**
+     * Returns a specified element from the Map object. If the value that is associated to the provided key is an object, then you will get a reference to that object and any change made to that object will effectively modify it inside the Map.
+     * @returns - Returns the element associated with the specified key. If no element is associated with the specified key, undefined is returned.
+     */
     async get(key: K): Promise<V | undefined> {
         await this.refresh();
         const entry = this.map.get(key);
@@ -52,6 +48,10 @@ export class Store<K extends string | number | symbol, V> {
         return undefined;
     }
 
+    /**
+     * @param key
+     * @returns boolean indicating whether an element with the specified key exists or not.
+     */
     async has(key: K): Promise<boolean> {
         await this.refresh();
         const entry = this.map.get(key);
@@ -65,6 +65,10 @@ export class Store<K extends string | number | symbol, V> {
         return false;
     }
 
+    /**
+     * @param key
+     * @returns true if an element in the Map existed and has been removed, or false if the element does not exist.
+     */
     delete(key: K): boolean {
         return this.map.delete(key);
     }
@@ -73,26 +77,42 @@ export class Store<K extends string | number | symbol, V> {
         this.map.clear();
     }
 
+    /**
+     * @returns the number of elements in the Map.
+     */
     size(): number {
         this.cleanUpExpiredEntries();
         return this.map.size;
     }
 
+    /**
+     * @returns an iterable of key, value pairs for every entry in the map.
+     */
     entries() {
         this.cleanUpExpiredEntries();
         return this.map.entries();
     }
 
+    /**
+     * @returns an iterable of values in the map
+     */
     values() {
         this.cleanUpExpiredEntries();
         return this.map.values();
     }
 
+    /**
+     * @returns an iterable of keys in the map
+     */
     keys() {
         this.cleanUpExpiredEntries();
         return this.map.keys();
     }
 
+    /**
+     * Executes a provided function once per each key/value pair in the Map, in insertion order.
+     * @param callback
+     */
     forEach(callback: (value: V, key: K) => void): void {
         this.cleanUpExpiredEntries();
         this.map.forEach((entry, key) => {
@@ -132,9 +152,9 @@ export class Store<K extends string | number | symbol, V> {
         if (!this.options) throw new Error("Options is needed.");
         return await deleteGithubFile({
             token: this.options.token,
-            repoOwner: this.options.owner,
-            repoName: this.options.repo,
-            filePath: this.options.path,
+            owner: this.options.owner,
+            repo: this.options.repo,
+            path: this.options.path,
             branch: this.options.branch,
         });
     }
@@ -156,9 +176,23 @@ export class Store<K extends string | number | symbol, V> {
         }, interval);
     }
 
+    private async refresh() {
+        if (this.map.size === 0 && this.options) {
+            const map = await getMap<K, V>({
+                token: this.options.token,
+                owner: this.options.owner,
+                repo: this.options.repo,
+                path: this.options.path,
+                branch: this.options.branch,
+            });
+            if (!map) return;
+            this.map = map;
+        }
+    }
+
     private async saveToGitHub(
         options: {
-            token: string;
+            token?: string;
             owner: string;
             repo: string;
             path: string;
@@ -166,16 +200,17 @@ export class Store<K extends string | number | symbol, V> {
         },
     ) {
         try {
+            if (!options.token) throw new Error("GITHUB_TOKEN is needed");
             if (this.options) {
                 const record = mapToRecord(this.map);
                 const data = JSON.stringify(record);
                 return await uploadFileToGitHub(
                     {
                         token: options.token,
-                        repoOwner: options.owner,
-                        repoName: options.repo,
-                        filePath: options.path,
-                        fileContent: data,
+                        owner: options.owner,
+                        repo: options.repo,
+                        path: options.path,
+                        content: data,
                         branch: options.branch,
                     },
                 );
@@ -214,19 +249,20 @@ function mapToRecord<K extends string | number | symbol, V>(
 
 async function getSHA(
     options: {
-        token: string;
-        repoOwner: string;
-        repoName: string;
-        filePath: string;
+        token?: string;
+        owner: string;
+        repo: string;
+        path: string;
         branch?: string;
     },
 ): Promise<string | undefined> {
+    if (!options.token) throw new Error("GITHUB_TOKEN is needed");
     const octokit = new Octokit({ auth: options.token });
     try {
         const response = await octokit.repos.getContent({
-            owner: options.repoOwner,
-            repo: options.repoName,
-            path: options.filePath,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             ref: options.branch,
         }) as any;
 
@@ -238,11 +274,11 @@ async function getSHA(
 
 async function uploadFileToGitHub(
     options: {
-        token: string;
-        repoOwner: string;
-        repoName: string;
-        filePath: string;
-        fileContent: string;
+        token?: string;
+        owner: string;
+        repo: string;
+        path: string;
+        content: string;
         branch?: string;
     },
 ) {
@@ -251,9 +287,9 @@ async function uploadFileToGitHub(
         const res = await getFileFromGithub(
             {
                 token: options.token,
-                repoOwner: options.repoOwner,
-                repoName: options.repoName,
-                filePath: options.filePath,
+                owner: options.owner,
+                repo: options.repo,
+                path: options.path,
                 branch: options.branch,
             },
         ) as any;
@@ -261,22 +297,22 @@ async function uploadFileToGitHub(
         const sha = res
             ? await getSHA({
                 token: options.token,
-                repoOwner: options.repoOwner,
-                repoName: options.repoName,
-                filePath: options.filePath,
+                owner: options.owner,
+                repo: options.repo,
+                path: options.path,
                 branch: options.branch,
             })
             : undefined;
 
         const message = res
-            ? `Update ${options.filePath}`
-            : `Create ${options.filePath}`;
+            ? `Update ${options.path}`
+            : `Create ${options.path}`;
         return await octokit.repos.createOrUpdateFileContents({
-            owner: options.repoOwner,
-            repo: options.repoName,
-            path: options.filePath,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             message,
-            content: btoa(options.fileContent),
+            content: btoa(options.content),
             sha,
             branch: options.branch,
         });
@@ -287,19 +323,20 @@ async function uploadFileToGitHub(
 
 async function getMap<K extends string | number | symbol, V>(
     options: {
-        token: string;
+        token?: string;
         owner: string;
         repo: string;
         path: string;
         branch?: string;
     },
 ) {
+    if (!options.token) throw new Error("GITHUB_TOKEN is needed");
     const rr = await getFileFromGithub(
         {
             token: options.token,
-            repoOwner: options.owner,
-            repoName: options.repo,
-            filePath: options.path,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             branch: options.branch,
         },
     ) as any;
@@ -309,18 +346,19 @@ async function getMap<K extends string | number | symbol, V>(
 }
 
 async function getFileFromGithub(options: {
-    token: string;
-    repoOwner: string;
-    repoName: string;
-    filePath: string;
+    token?: string;
+    owner: string;
+    repo: string;
+    path: string;
     branch?: string;
 }) {
+    if (!options.token) throw new Error("GITHUB_TOKEN is needed");
     const octokit = new Octokit({ auth: options.token });
     try {
         const res = await octokit.repos.getContent({
-            owner: options.repoOwner,
-            repo: options.repoName,
-            path: options.filePath,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             ref: options.branch,
         });
         return res.data;
@@ -331,30 +369,31 @@ async function getFileFromGithub(options: {
 
 async function deleteGithubFile(
     options: {
-        token: string;
-        repoOwner: string;
-        repoName: string;
-        filePath: string;
+        token?: string;
+        owner: string;
+        repo: string;
+        path: string;
         branch?: string;
     },
 ) {
+    if (!options.token) throw new Error("GITHUB_TOKEN is needed");
     const octokit = new Octokit({ auth: options.token });
     try {
         const sha = await getSHA({
             token: options.token,
-            repoOwner: options.repoOwner,
-            repoName: options.repoName,
-            filePath: options.filePath,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             branch: options.branch,
         });
         if (!sha) throw new Error("SHA is needed");
         const res = await octokit.repos.deleteFile({
-            owner: options.repoOwner,
-            repo: options.repoName,
-            path: options.filePath,
+            owner: options.owner,
+            repo: options.repo,
+            path: options.path,
             sha,
             branch: options.branch,
-            message: `Delete ${options.filePath}`,
+            message: `Delete ${options.path}`,
         });
         return res.data;
     } catch (error) {
